requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pillow>=9.5.0

# Object detection
pycocotools>=2.0.6

# Training utilities
tqdm>=4.65.0
pyyaml>=6.0

# Optional but recommended
tensorboard>=2.13.0
matplotlib>=3.7.0
opencv-python>=4.8.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

